Move `run_in_transaction` and related tests
It's not a context manager anymore!
